refactor(web): merge blocks paginated table utils in blocks view (#452)
