--- apps/web/src/components/PaginatedTable/helpers/index.ts
@@ -1,2 +1,0 @@
-export { getBlobsTableRows, blobsTableHeaders } from "./blobs";
-export { getBlocksTableRows, blocksTableHeaders } from "./blocks";
